+++ .idea/.idea.AcmeStudioRefactorTest_IF/.idea/.name
@@ -1,0 +1,1 @@
+AcmeStudioRefactorTest_IF
