--- AcmeStudios.ApiRefactor/AutoMapperProfile.cs
@@ -1,15 +1,0 @@
-﻿using AcemStudios.ApiRefactor.DTOs;
-using AutoMapper;
-
-namespace AcemStudios.ApiRefactor
-{
-    public class AutoMapperProfile : Profile
-    {
-        public AutoMapperProfile()
-        {
-            CreateMap<StudioItem, GetStudioItemDto>();
-            CreateMap<AddStudioItemDto, StudioItem>();
-            CreateMap<StudioItem, GetStudioItemHeaderDto>();
-        }
-    }
-}
