+++ AcmeStudios.ApiRefactor/DTOs/GetStudioItemHeaderDto.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace AcemStudios.ApiRefactor.DTOs
+namespace AcmeStudios.ApiRefactor.DTOs
 {
     public class GetStudioItemHeaderDto
     {
