+++ AcmeStudios.ApiRefactor/Program.cs
@@ -1,18 +1,17 @@
+using AcmeStudios.ApiRefactor.Configuration;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Hosting;
 
-namespace AcemStudios.ApiRefactor
+namespace AcmeStudios.ApiRefactor
 {
     public class Program
     {
         public static void Main(string[] args)
         {
             CreateHostBuilder(args).Build().Run();
-
-
         }
 
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
+        private static IHostBuilder CreateHostBuilder(string[] args) =>
             Host.CreateDefaultBuilder(args)
                 .ConfigureWebHostDefaults(webBuilder =>
                 {
