--- AcmeStudios.ApiRefactor/Startup.cs
@@ -1,61 +1,0 @@
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using AutoMapper;
-
-namespace AcemStudios.ApiRefactor
-{
-    public class Startup
-    {
-        public Startup(IConfiguration configuration)
-        {
-            Configuration = configuration;
-        }
-
-        public IConfiguration Configuration { get; }
-
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddCors(options =>
-             {
-                 options.AddPolicy("AllowMyOrigin",
-                 builder => builder.WithOrigins("http://localhost:4200")
-                 .AllowAnyHeader()
-                 .AllowAnyMethod()
-                 );
-             });
-
-            services.AddControllers();
-
-            services.AddSwaggerGen();
-
-            services.AddAutoMapper(typeof(Startup));
-        }
-
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-        {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();                
-            }
-
-            app.UseStaticFiles();
-
-            app.UseHttpsRedirection();
-
-            app.UseRouting();
-
-            app.UseAuthorization();
-
-            app.UseCors("AllowMyOrigin");
-
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapControllers();
-                endpoints.MapControllers().RequireCors("AllowMyOrigin");
-            });
-        }
-    }
-}
